+++ src/MelloSilveiraTools/ExtensionMethods/NpgsqlCommandExtensions.cs
@@ -30,7 +30,7 @@ public static class NpgsqlCommandExtensions
             if (attribute != null)
             {
                 object? value = property.GetValue(entity);
-                command.Parameters.AddWithValue(property.Name, GetDbTypeFromPropertyType(property.PropertyType), value ?? DBNull.Value);
+                command.Parameters.AddWithValue(property.Name, property.PropertyType.GetDbTypeFromPropertyType(), value ?? DBNull.Value);
             }
         }
 
@@ -38,18 +38,22 @@ public static class NpgsqlCommandExtensions
     }
 
     /// <summary>
-    /// Returns the <see cref="NpgsqlDbType"/> from property type.
+    /// Sets the parameter for sql command.
     /// </summary>
-    /// <param name="type"></param>
+    /// <typeparam name="TEntity"></typeparam>
+    /// <param name="command"></param>
+    /// <param name="parameters"></param>
     /// <returns></returns>
-    /// <exception cref="Exception"></exception>
-    private static NpgsqlDbType GetDbTypeFromPropertyType(Type type)
+    public static NpgsqlCommand SetCommandParameters<TEntity>(this NpgsqlCommand command, IEnumerable<NpgsqlParameter> parameters)
     {
-        if (type == typeof(string)) return NpgsqlDbType.Text;
-        if (type == typeof(double)) return NpgsqlDbType.Double;
-        if (type == typeof(long)) return NpgsqlDbType.Bigint;
-        if (type == typeof(DateTime) || type == typeof(DateTimeOffset)) return NpgsqlDbType.Timestamp;
-        if (type == typeof(IList) || type == typeof(IEnumerable) || type == typeof(IEnumerator)) return NpgsqlDbType.Array;
-        throw new ArgumentOutOfRangeException(nameof(type), $"Invalid type: '{type.FullName}'.");
+        if (parameters.IsNullOrEmpty())
+            return command;
+
+        foreach (NpgsqlParameter parameter in parameters)
+        {
+            command.Parameters.Add(parameter);
+        }
+
+        return command;
     }
 }