+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Resilience pipeline for default cases and PostgreSQL repositories.
 - Logger.
 - GetDeleteByPrimaryKeyQuery, GetSelectByPrimaryKeyQuery and GetUpdateByPrimaryKeyQuery methods on ISqlProvider.
+- BuildParameters method on TypeExtensions.
 
 ## [1.0.3] - 2025-06-11
 ### Added