+++ CHANGELOG.md
@@ -6,13 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## UNRELEASED
 ### Added
-- Resilience pipeline for default cases and PostgreSQL repositories.
 - Logger.
 - GetDeleteByPrimaryKeyQuery, GetSelectByPrimaryKeyQuery and GetUpdateByPrimaryKeyQuery methods on ISqlProvider.
-- Extension methods evolving NPGSQL classes and IFormFile.
+- Extension methods evolving NPGSQL classes, IFormFile, Type and string.
 - JWE authentication.
 - Methods to enable Swagger using v1.
 - Service for encryption.
+- String builder using pooled buffers and stack allocation optimizations.
+- Resilience pipeline for default cases and PostgreSQL repositories.
+- Basic and PostgreSQL implementations for database repository.
+- Basic implementations for use cases operations.
+- Basic implementations for API service agents.
 ### Removed
 - ReferencedPropertyName property from ForeignKeyColumnAttribute.