+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## [1.1.0] - 205-11-29
 ### Added
 - Logger.
 - GetDeleteByPrimaryKeyQuery, GetSelectByPrimaryKeyQuery and GetUpdateByPrimaryKeyQuery methods on ISqlProvider.
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Basic and PostgreSQL implementations for database repository.
 - Basic implementations for use cases operations.
 - Basic implementations for API service agents.
+- Mechanical of materials.
 ### Removed
 - ReferencedPropertyName property from ForeignKeyColumnAttribute.